Rename deployment container
