+++ packages/client/src/ee/main.tsx
@@ -22,7 +22,7 @@ import ScheduleCreatePage from 'ee/containers/scheduleCreatePage';
 import ScheduleListContainer from 'ee/containers/scheduleList';
 import ModelListContainer from 'ee/containers/modelList';
 import ModelDetailContainer from 'ee/containers/modelDetail';
-import ModelDeploymentListContainer from 'ee/containers/modelDeploymentList';
+import DeploymentListContainer from 'ee/containers/deploymentList';
 import DeploymentDetailContainer from 'ee/containers/deploymentDetail';
 import DeploymentCreatePage from 'ee/containers/deploymentCreatePage';
 import DeploymentEditPage from 'ee/containers/deploymentEditPage';
@@ -123,7 +123,7 @@ class Main extends React.Component {
 
             {/* Model Deployment */}
             <Route path={`${appPrefix}g/:groupName/deployments`} exact>
-              <ListContainer Com={ModelDeploymentListContainer} />
+              <ListContainer Com={DeploymentListContainer} />
             </Route>
             <Route path={`${appPrefix}g/:groupName/deployments/create`} exact>
               <DeploymentCreatePage />
